Descriptor: modify for some bugs details: 1. delete trackAllocations_ 2. modify size_t to int64 3. modify IntegerRef to NumberRef issue: https://gitee.com/openharmony/ark_js_runtime/issues/I5BCAD

+++ ecmascript/tooling/base/pt_events.h
@@ -892,12 +892,12 @@ public:
         return "Profile.PreciseCoverageDeltaUpdate";
     }
 
-    size_t GetTimestamp() const
+    int64_t GetTimestamp() const
     {
         return timestamp_;
     }
 
-    PreciseCoverageDeltaUpdate &SetTimestamp(size_t timestamp)
+    PreciseCoverageDeltaUpdate &SetTimestamp(int64_t timestamp)
     {
         timestamp_ = timestamp;
         return *this;
